+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Cache autocomplete options if they are fetched remotely
+
 ## [0.5.1] - 2024-05-14
 
 ### Fixed