+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
+- Return early if panel is not present in `PopoverComponent`
 - Fixes an issue where popover was breaking Chrome 133
 
 ## [0.5.1] - 2024-05-14