+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Remote autocomplete should not remove the `loading` attribute on `AbortError`
+
 ## [0.6.0] - 2025-02-07
 
 ### Added