+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.7.1] - 2025-07-04
+
+### Fixed
+
+- Fixed SCSS compile issues when using the `min` CSS function ([#165](https://github.com/Ambiki/impulse_view_components/pull/165))
+
 ## [0.7.0] - 2025-07-04
 
 ### Added
@@ -161,7 +167,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Everything!
 
-[unreleased]: https://github.com/Ambiki/impulse_view_components/compare/v0.7.0...HEAD
+[unreleased]: https://github.com/Ambiki/impulse_view_components/compare/v0.7.1...HEAD
+[0.7.1]: https://github.com/Ambiki/impulse_view_components/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/Ambiki/impulse_view_components/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/Ambiki/impulse_view_components/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/Ambiki/impulse_view_components/compare/v0.5.0...v0.5.1